requirements.txt
build==0.10.0
click==8.1.6
contourpy==1.0.5
cycler==0.11.0
fonttools==4.25.0
h5py==3.9.0
imageio==2.31.1
kiwisolver==1.4.4
lazy_loader==0.3
matplotlib==3.7.1
munkres==1.1.4
networkx==3.1
nibabel==5.1.0
numpy==1.25.0
packaging==23.0
Pillow==9.4.0
pip==23.1.2
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
PyWavelets==1.4.1
runstats==2.0.0
scikit-image==0.21.0
scipy==1.11.1
setuptools==67.8.0
six==1.16.0
tifffile==2023.7.18
tornado==6.2
tqdm==4.65.0
wheel==0.38.4
